moved root and dispatch types to the appTypes file
