+++ src/types/appTypes.ts
@@ -1,4 +1,9 @@
 import { Session } from '@supabase/supabase-js';
+import store from '../store/store';
+
+// Infer the `RootState` and `AppDispatch` types from the store itself
+export type RootState = ReturnType<typeof store.getState>;
+export type AppDispatch = typeof store.dispatch;
 
 export type ChannelItemData = {
   name: string;
